refactoring changes naming and reuses functions
